Fix #2: Move VuexValidator code to own file; Use lodash-es
